Problem: can't use anonymous S3 credentials
When trying to use S3 cache with anonymous credentials (for example, for importing publicly available layers), the cache is not used.

Solution: enable anonymous credentials

According to https://docs.aws.amazon.com/sdk-for-go/api/aws/credentials/:

```
var AnonymousCredentials = NewStaticCredentials("", "", "")
```
"AnonymousCredentials is an empty Credential object that can be used as dummy placeholder credentials for requests that do not need signed."

This change allows setting these empty credentials so that the client can be used anonymously.

Signed-off-by: Yurii Rashkovskii <yrashk@gmail.com>